redis: URL support for connection parameters
Make it possible to specify the Redis connection parameters via URL
which can be supplied in the config file with key "url" (and leaving
"host", "port", "password" and "db" empty/unspecified) or via
environment variable PASSARI_WORKFLOW_REDIS_URL or REDIS_URL.

The REDIS_URL environment variable is also used by the RQ workers by
default so specifying it will also make RQ work without a config file.
The default value if no connection parameters are given is to use db 0
on localhost (which is the same as RQ uses).

Additionally provide a function "get_redis_url" so that passari-web-ui
can use it to just get the URL without creating a Redis connection.

Also tune the PostgreSQL URL handling to support similar environment
variable based configuration via PASSARI_WORKFLOW_DB_URL or DATABASE_URL
variables and fix the password escaping to use correct `quote` function
instead of `quote_plus`.

Note: The get_redis_connection used to accept a db index argument, but
not anymore.  Implementing override of the db index for the URLs would
need some extra code, because the db index can be specified in various
formats in the URL, but the argument was not used by any callers (in
this code base or in passari-web-ui) so it's not worth it to implement.