Provide pas-db-migrate command and move migrations
Move alembic migration files to passari_workflow/db/migrations and
alembic.ini to its parent directory so that they are installed with the
package.

Add new script named "pas-db-migrate" which can be used to run the
alembic migration commands.  Update the docs accordingly.

Update the test_migrate to use the new path to alembic.ini.

This allows running the migrations even when installing from a wheel and
no source directory is available.